+++ src/ULinkGame.Server/Actors/IActorDirectoryCache.cs
@@ -1,0 +1,12 @@
+using ULinkGame.Cluster;
+
+namespace ULinkGame.Server.Actors;
+
+public interface IActorDirectoryCache
+{
+    bool TryGet(ActorId actorId, out NodeId node);
+
+    void Set(ActorId actorId, NodeId node);
+
+    void Remove(ActorId actorId);
+}